fix: Move Calculator tabs to its own file
