+++ src/app/energyCalculator/tabs/byPrice.tsx
